ステージ選択方法を変更した #276
