--- Assets/Scripts/StageSelect/SelectButtonController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ef66241689a612b4e9c715f687e273c6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
